Create MUSIC.py
MUSIC.py implements a High-resolution spectral estimation method of audio source localization. In particular, it implements the Multiple Signal Classification algorithm, which works on the concept of eigen value decomposition of the cross-spectrum correlation matrix.

This code uses the Standard Eigen Value Decomposition (SEVD) technique to decompose the averaged and stable correlation matrix into the Signal subspace and the Noise subspace. Since the magnitude of the eigen values corresponds directly to the power of the audio signal, a larger eigen value magnitude corresponds to the signal subspace. Using this information, a computation of the MUSIC power spectrum is done. The direction of arrival (DOA) can be determined by searching for the local peaks in the MUSIC power spectrum.